#13111 Testing. Complete, fix bug, and create default seed

+++ app/Http/Controllers/Backend/UserController.php
@@ -2,15 +2,32 @@
 
 namespace App\Http\Controllers\Backend;
 
+use App\Http\Requests\DeleteUserInformationRequest;
 use App\Http\Requests\UserSaveRequest;
 use App\Http\Requests\UserUpdateRequest;
+use App\Services\UserService;
 use App\User;
 use App\Http\Controllers\Controller;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
 {
+    /**
+     * @var UserService
+     */
+    private $userService;
+
+    /**
+     * UserController constructor.
+     * @param UserService $userService
+     */
+    public function __construct(UserService $userService)
+    {
+        $this->userService = $userService;
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -123,6 +140,12 @@ class UserController extends Controller
     {
         DB::beginTransaction();
         try {
+            $user->load('articles', 'favorite');
+
+            if (count($user->articles) > 0 or count($user->favorite) > 0)
+            {
+                return redirect()->route('confirmation.delete', ['user' => $user->id])->with('warning', 'The user has articles. Are you sure you want to erase all data linking to it? Operation is not possible to roll back!');
+            }
             $user->delete();
 
             DB::commit();
@@ -133,6 +156,40 @@ class UserController extends Controller
             DB::rollback();
 
             return back()->with('error', 'Error!' . $e);
+        }
+    }
+
+    /**
+     * @param $id
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     */
+    public function confirmationDelete($id)
+    {
+        return view('backend.dashboard.user.confirmation_delete',[
+            'user' => User::find($id),
+            'title' => 'Delete ',
+            'message' => session('warning')
+        ]);
+    }
+
+    /**
+     * @param DeleteUserInformationRequest $request
+     * @return \Illuminate\Http\RedirectResponse
+     */
+    public function deleteAllInformationForUser(DeleteUserInformationRequest $request)
+    {
+        DB::beginTransaction();
+        try {
+            $message = $this->userService->deleteAllInformationForUser($request->except('_token'));
+
+            DB::commit();
+
+            return redirect()->route('users.index')->with('success', $message);
+
+        } catch (\Throwable $e) {
+            DB::rollback();
+
+            return back()->with('error', 'Error! Not found!');
         }
     }
 }
